`NewIncomingStream` notifications now contain a stream name.
Stream name will be needed where we need wildcard publishing capabilities.  In order to serve video with RTMP playback capabilities when a wildcard stream key is given, we need the stream name to know what stream key to serve which video on.  Likewise, for archival purposes (like generating HLS feeds) we need a stream name to know how to name the files when multiple streams are coming through the same workflow.