Track which steps are still waiting to be instantiated.
Previously workflows would throw an error when checking if all steps are active because the workflow did not yet have an instance of the step to check the status.  This commit now allows us to differentiate between an invalid step id (one that is unexpected due to a bug) and a step id that we know about but we haven't received an instance yet.