Send a signal to a workflow to shut it down.
I was relying on the workflow manager dropping the sender to trigger a shutdown signal (due to all the senders being gone).  However, I had forgotten that I was storing a clone of the sender in a future, so the workflow manager would get notified if the workflow's actor died.  This kept at least one sender alive, which prevented the actor from collapsing as expected.

Now workflows support being manually requested to stop, and the workflow manager will trigger this when it receives a request to remove that specific workflow.