Reliable rtmp registration cleanup
If a connection is active when the publish/watch registration is removed, the connection would persist afterwards.  This commit cleans up the process so when a publish or watch registration is removed all related connections are disconnected as well.

This also cleans up the rtmp application from memory if no registrations are active for it anymore.