Fixed bug where sequence headers could be missed
In the ffmpeg based steps, since registration happens *after* a step becomes active it's possible that sequence headers are received by the step before we have registered for RTMP watchers.  We were putting these early media packets into a pending collection, but we never did anything with them.  This caused sequence headers to sometimes be missed, and thus cause video to not be able to be decoded by watchers.

This was fixed by sending off any pending messages when watch registration becomes valid.