Maintain ffmpeg id on ffmpeg failure
When ffmpeg stops unexpectedly (usually due to an error or a disconnection) ffmpeg will automatically be restarted.  This was causing the new instance of ffmpeg to have a brand new ffmpeg id.

This wasn't optimal as it would cause a spam of logs if ffmpeg was experiencing an error (for instance the hls directory doesn't exist).  This will also make it harder for users to hunt down issues at runtime, as each run will have logs spread all around the place.

So re-using the ffmpeg should make for an easier debugging workflow.