Added Reactor functionality
Reactors are a primary mechanism for dynamic workflows in mmids. When they receive a query for a stream name, the reactor will reach out to a configured external service to get the definition of the workflow that is tied to that stream name. It then has the workflow manager create the returned workflow and returns the name of the workflow to the original requestor.

Endpoints use these reactors to know if a stream name is allowed, allowing fine control over which clients are allowed to connect.

Workflow steps use reactors in order to dynamically forward media to different workflows based on the stream names.