Fully unwrap FLV video packets
Previously we were only wrapping/unwrapping the FLV tag, but leaving the `AVCVIDEOPACKET` with the rest of the h264 data.  This causes parsing errors downstream.

So the code now takes out the composition time offset from the RTMP packet, and uses that to set the pts and dts values that get passed around by mmids.