Rtmp endpoint shouldn't die if socket listener stops
The Rtmp endpoint was waiting for `TcpSocketResponse` messages to come back from the socket manager, and if the channel ended up being closed it mistakenly assumed the socket manager as a whole was closed.  This was incorrect as that channel is specific to that one specific port, and thus the socket manager may still be available for other ports.

So instead of shutting the Rtmp endpoint down when the response channel shuts down, it only removes that single port, thus keeping the Rtmp endpoint alive for other ports it may be active on (or to serve new requests to become active).

Fixing this exposed that workflow steps using RTMP were not being notified when their Rtmp endpoint registration has been removed.  This was caused because `wait_for_publisher_channel_closed` and `wait_for_watcher_channel_closed` were keeping the channel alive by waiting until it saw the channel was closed.

Therefore, a cancellation token was added as part of the registration structures that is now also used to notify that a registrant is gone if either the channel is gone or the cancellation token channel is gone.  This now allows workflow steps to enter an error state if the socket manager has a port closed (or listener actor fails to load).

Fixes #33