Flexible payload types (#37)
Currently the `MediaNotificationContent` enumeration has several issues preventing wider use:

1. They have two variants, one for audio and one for video, which makes it hard to route non-AV media payloads through the workflow engine
2. The audio and video variants have specific per-packet metadata as properties, which may not be relevant for all audio or video payloads that flow through the system (for example, not all audio codecs have sequence headers).  Similarly, there are times when consumers of mmids would want to be able to add new metadata information on a per packet basis, and the current structure doesn't allow that.
3. The audio and video codecs are currently defined as a closed enumeration, which makes it impossible for other consumers to use a codec the core library doesn't have an entry for.

To fix all of these a new variant was created, the `MediaNotificationContent::MediaPayload`.  This allows a generic payload to be produced with a flexible set of per-packet metadata, and an `Arc<String>` for codecs to allow consumers to define their own.

The per-packet metadata collection has been designed to reduce allocations required as the application runs, by allowing systems that create media notifications to maintain an arena style memory buffer that the metadata collections can be used.  This should cause allocations to be heavy at the start, but start to slow down and be amortized as the arena starts having more free space to put new metadata collections into it.

The Audio and Video variants will be removed in a follow-up PR.