Remove actor notification boilerplate (#47)
A lot of actors were spinning up a bunch of tokio tasks + tokio selects just to facilitate notifications back to the actor. This PR consolidates a lot of the boilerplate into actor utility functions that are easier to manage and harder to cause issues (such as forgetting to break out of a loop on an actor closing).