Workflow steps should send futures via channels (#49)
Previously, workflow steps would box up their futures and return them to the workflow runner for execution. This caused excessive allocations for every future and added extra steps to getting that future to run.

Instead, the API has been refactored to allow the workflow runner to pass workflow steps a channel it can send the results of the step's futures to. This allows workflow steps to spawn their own tasks for their own futures, and only involve the workflow runner when a result has been computed.

This also helps the ergonomics of writing workflow steps, as each future result doesn't need to pass back the receiver to restart the future. Instead they can continually recv from a receiver in a single future until the channel is closed.