Create workflow step futures media channel (#50)
When a workflow step receives media from an endpoint, it always gets each media notification from a future, usually from a tokio channel. This means each media payload comes back to the workflow step as a `Box<dyn StepFutureResult>`, and needs to be unpacked and manually placed into the workflow's `StepOutputs.media` vector by the step.  

Not only does this cause extra steps when creating workflow steps, but it also means that *every* media packet must be boxed, causing unnecessary heap copying and allocations.

This PR updates the workflow step futures channel to allow transferring strongly typed media packets as part of a completed future. This allows a workflow step to designate that a future returns media packets destined for the next step, thus removing the need for boxing and removing the need for workflow steps to manually move media packets around to get them to the next step.  

Note that while this is implemented from an API standpoint, and verified via workflow runner tests, this isn't utilized by most of the built in workflow steps (except the gstreamer transcoding one). This is because most workflow steps built in utilize the RTMP endpoint, which needs to be refactored to take advantage of this new pattern (and that's out of scope for this PR, but will come later).