Redesign workflow step traits (#52)
The workflow step trait was previously designed to be queryable even when it entered an error or shutdown state.  This caused a lot of boilerplate with simple steps, and made it easy to accidentally leak resources by not handling a shutdown incorrectly.  It also gave uncertainty that execute might accidentally be called when it's in the wrong state.

This is now fixed so that the aspects we might query about a workflow step (its definition and it's status specifically) are separate. When a step is created or executed it returns a status, and if that status is an error state then it gets shut down.

Shutting down a workflow step is now just a matter of dropping it, which makes it much easier to automatically clean up after itself once a step is no longer being used (intentionally or not).